==> notebooks/01_LoadData_Python.ipynb <==
### Load Data from remote server

#### Output
* winequality-red.csv

#### Azure blob

import os
from azure.storage.blob import BlockBlobService

# azure credentials
# make sure the credentials are saved as environment variables.
connect_str = os.getenv('AZURE_STORAGE_CONNECTION_STRING')

# file and directory info
container_name = 'datascience-data' 
file_name = 'winequality-red.csv'
project_dir = os.path.dirname(os.getcwd())
final_file_dir = os.path.join(project_dir, 'data/raw', file_name)

# for azure.storage.blob 2.1.0
blob = BlockBlobService(connection_string=connect_str)
blob.get_blob_to_path(container_name, 
                      file_name, 
                      final_file_dir)

# # for azure.storage.blob 12.0.2
# blob = BlobClient.from_connection_string(conn_str=connect_str, 
#                                          container_name=container_name, 
#                                          blob_name=file_name)
# # save the file to the assigned directory
# final_file_dir = os.path.join(project_dir, 'data/raw', file_name)
# with open(final_file_dir, "wb") as my_blob:
#     blob_data = blob.download_blob()
#     blob_data.readinto(my_blob)

#### S3 bucket 

# import os
# import boto3

# # bucket credentials
# # make sure the credentials are saved as environment variables.
# access_key_id = os.environ['AWS_ACCESS_KEY_ID']
# secret_access_key = os.environ['AWS_SECRET_ACCESS_KEY']
# 
# # file and directory info
# bucket_name = 'choux130' 
# file_name = 'winequality-red.csv'
# project_dir = '/Users/chou/Desktop/mlflow_dvc_cookiecutter/DSProjectTemplate'
# 
# s3 = boto3.client('s3', 
#                   aws_access_key_id = access_key_id, 
#                   aws_secret_access_key = secret_access_key)
# 
# # save the file to the assigned directory
# final_file_dir = os.path.join(project_dir, 'data/raw', file_name)
# s3.download_file(bucket_name, 'DS_project_data/' + file_name, final_file_dir)

#### URL (Use Github as an example)

# import pandas as pd

# csv_url = 'https://raw.githubusercontent.com/choux130/Wine-Quality-Dataset/master/winequality-red.csv'
# data = pd.read_csv(csv_url, error_bad_lines = False)

# # file and directory info
# file_name = 'winequality-red.csv'
# project_dir = '/Users/chou/Desktop/mlflow_dvc_cookiecutter/DSProjectTemplate'

# # save the file to the assigned directory
# final_file_dir = os.path.join(project_dir, 'data/raw', file_name)
# data.to_csv(final_file_dir, index=None, header=True, float_format='%g')
