==> notebooks/02_SplitData_R.ipynb <==
### Split Data into Train and Test dataset

#### Input 
* winequality-red.csv

#### Output
* train.parquet
* test.parquet

# params
random_seed = 123
percent_train = 0.8

suppressMessages(library(dplyr))

# file and directory info
file_name = 'winequality-red.csv'
project_dir = dirname(getwd())
raw_folder = "/data/raw/"
interim_folder = "/data/interim/"

# input
raw_data = data.table::fread(paste0(project_dir, raw_folder, file_name), fill = TRUE) %>%
dplyr::as_tibble(.)

# functions
SplitData = function(raw_data, random_seed, percent_train){
    set.seed(random_seed)
    
    train_index = sample(1:nrow(raw_data), percent_train * nrow(raw_data))
    test_index = setdiff(1:nrow(raw_data), train_index)

    train = raw_data[train_index, ]
    test = raw_data[test_index, ]
    
    return(list(train = train, test = test))
}

# ouptut and save
train_test = SplitData(raw_data, random_seed, percent_train)
arrow::write_parquet(train_test$train, paste0(project_dir, interim_folder, "train.parquet"))
arrow::write_parquet(train_test$test, paste0(project_dir, interim_folder, "test.parquet"))

dim(raw_data)
head(raw_data)

lapply(train_test, dim)
lapply(train_test, head)
