==> notebooks/06_Predict_pyfunc.ipynb <==
### Predict

import mlflow.pyfunc
import os 
import pandas as pd

# load the model by passing the model path
run_id = 'f98a86efdc1441e6ba868509605a6bb8'
mlflow_pyfunc_model_path = os.getcwd() + '/mlruns/1/' + run_id + '/artifacts/model'
# mlflow_pyfunc_model_path = 'wasbs://datascience-mlflow-artifacts@ting.blob.core.windows.net/3/7e8c41558a3e429882ecff3b46e3ff39/artifacts/model'
loaded_model = mlflow.pyfunc.load_model(mlflow_pyfunc_model_path)

# file and directory info
project_dir = os.path.dirname(os.getcwd())
interim_folder = '/data/interim/'

# take test dataset as pred
model_input = pd.read_parquet(project_dir + interim_folder + 'test.parquet')
# predict
pred = loaded_model.predict(model_input)
pred.tolist()[0:10]

# # serve the model locally with port 1234
# mlflow models serve -m <mlflow_pyfunc_model_path> -p 1234
# # get prediction using POST
# curl http://127.0.0.1:1234/invocations -H "Content-Type:application/json; format=pandas-split" --data '{"columns":["alcohol", "chlorides", "citric acid", "density", "fixed acidity", "free sulfur dioxide", "pH", "residual sugar", "sulphates", "total sulfur dioxide", "volatile acidity"],"data":[[1,1, 1, 1, 1, 1, 1, 1, 1, 1, 1]]}'
